// src/navigation/types/ScreenName.ts
export enum ScreenName {
  connection = 'ConnectionScreen',
  camera = 'CameraScreen',
  connectionApproval = 'ConnectionApprovalScreen',
  createWallet = 'CreateWalletScreen',
  unlockWallet = 'UnlockWalletScreen',
  restoreWallet = 'RestoreWallet',
  dashboard = 'Dashboard',
  sendEth = 'SendEth',
  sendEthBriefing = 'SendEthBriefing',
  sendEthCompleted = 'SendEthCompleted',
  splash = 'SplashScreen',
  receiveEth = 'ReceiveEthScreen',
}